Introduce GitIgnore.js, fix some glitches
